style changes + toast modified
