+++ package.json
@@ -13,6 +13,7 @@
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "react-toastify": "^9.1.3",
     "sass": "^1.62.1",
     "web-vitals": "^2.1.4"
   },
