+++ src/components/Header/index.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Header = () => {
+  return <h1>Budget planner</h1>
+}
+
+export default Header
